test: separate offline tests from live-network integration tests
Split mixed test files so every test that hits a live host (httpbin.org et al.)
lives in a *IntegrationTests class, leaving a deterministic offline suite
(85 tests). CI now runs 'swift test --skip IntegrationTests'; integration
tests remain runnable locally via the full 'swift test'.
